refactor(beads): move description parsing under lib
- move the extracted description-field parser into atelier.lib.beads so it stays backend-scoped during the facade drain
- update the legacy facade, beads-backed store, promote-epic skill, and parity tests to import the private helper from its new home
- preserve the existing read/discovery behavior while removing the new top-level module surface introduced in the earlier slice
