+++ tests/atelier/test_prefix_migration_convergence.py
@@ -9,9 +9,9 @@ from unittest.mock import patch
 
 import pytest
 
-import atelier.bead_description_fields as bead_fields
 import atelier.prefix_migration_drift as prefix_migration_drift
 import atelier.worktrees as worktrees
+from atelier.lib.beads import description_fields as bead_fields
 from atelier.worker.session import worktree as session_worktree
 
 doctor_cmd = importlib.import_module("atelier.commands.doctor")
